Reorganize lessons and remove Setup task
