+++ learning/katas/README.md
@@ -23,4 +23,4 @@ that can help you to learn Apache Beam concepts and programming model hands-on.
 Built based on [JetBrains Educational Products](https://www.jetbrains.com/education/), Beam Katas 
 objective is to provide a series of structured hands-on learning experiences for learners 
 to understand about Apache Beam and its SDKs by solving exercises with gradually increasing 
-complexity. Beam Katas are available for Java, Python, and Go (In Development) SDKs.
+complexity. Beam Katas are available for Java, Python, and Go SDKs.
