+++ learning/katas/go/Introduction/Hello Beam/Hello Beam Test/task-info.yaml
@@ -19,20 +19,18 @@
 
 type: edu
 files:
-- name: test/task_test.go
-  visible: false
-- name: task.go
-  visible: true
-  placeholders:
-  - offset: 961
-    length: 26
-    placeholder_text: 'TODO: create a new pipeline'
-  - offset: 1012
-    length: 34
-    placeholder_text: 'TODO: execute the pipeline'
 - name: go.mod
   visible: false
+- name: pkg/task/task.go
+  visible: true
 - name: go.sum
   visible: false
-- name: check.go
-  visible: false
+- name: test/task_test.go
+  visible: true
+  placeholders:
+  - offset: 1119
+    length: 36
+    placeholder_text: TODO()
+  - offset: 1165
+    length: 12
+    placeholder_text: TODO()
