+++ learning/katas/go/Introduction/Hello Beam/lesson-info.yaml
@@ -1,4 +1,3 @@
-
 #
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
@@ -19,4 +18,5 @@
 #
 
 content:
-  - Hello Beam
+- Hello Beam
+- Hello Beam Test
