--- learning/katas/go/Introduction/Hello Test/Hello Test/task-info.yaml
@@ -1,39 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#  http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-type: edu
-files:
-- name: test/task_test.go
-  visible: true
-  placeholders:
-  - offset: 1067
-    length: 22
-    placeholder_text: 'TODO: provide a value to validate the test.'
-  - offset: 1124
-    length: 31
-    placeholder_text: 'TODO: assert got == want here.'
-  - offset: 1165
-    length: 12
-    placeholder_text: 'TODO: run the test.'
-- name: test/output.txt
-  visible: false
-- name: go.mod
-  visible: false
-- name: go.sum
-  visible: false
