--- learning/katas/go/Introduction/Setup/Setup/task.md
@@ -1,51 +1,0 @@
-<!--
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
--->
-
-# Welcome to Apache Beam!
-
-Apache Beam is an open source, unified model for defining both batch and streaming data-parallel
-processing pipelines. Using one of the open source Beam SDKs, you build a program that defines the
-pipeline. The pipeline is then executed by one of Beam’s supported distributed processing
-back-ends, which include Apache Apex, Apache Flink, Apache Spark, and Google Cloud Dataflow.
-
-Beam is particularly useful for Embarrassingly Parallel data processing tasks, in which the
-problem can be decomposed into many smaller bundles of data that can be processed independently
-and in parallel. You can also use Beam for Extract, Transform, and Load (ETL) tasks and pure data
-integration. These tasks are useful for moving data between different storage media and data
-sources, transforming data into a more desirable format, or loading data onto a new system.
-
-To learn more about Apache Beam, refer to [Apache Beam Overview](https://beam.apache.org/get-started/beam-overview/).
-
-This series of katas assumes some familiarity with the Go language and is not meant to teach you the language.
-We encourage you to head on over to [https://golang.org/doc/#learning](https://golang.org/doc/#learning) if you need
-a resource to get started with the Go language.
-
-**Kata:** Your first kata is to setup your first pipeline and run it on your machine.
-
-<div class="hint">
-  The Beam Go SDK provides a convenient starter function to create a new empty pipeline and its root scope: 
-  <a href="https://godoc.org/github.com/apache/beam/sdks/go/pkg/beam#NewPipelineWithRoot">
-  NewPipelineWithRoot</a>.
-</div>
-
-<div class="hint">
-  <a href="https://godoc.org/github.com/apache/beam/sdks/go/pkg/beam/x/beamx#Run">beamx.Run</a>
-  from the package beamx is a convenience function that defaults to the direct runner, 
-  but all beam-distributed runners and textio filesystems are implicitly registered.
-</div>
