+++ learning/katas/go/Introduction/section-info.yaml
@@ -18,6 +18,4 @@
 #
 
 content:
-- Setup
 - Hello Beam
-- Hello Test
